Creates an 'api' folder and creates a new 'server.js' file inside that folder to separate our server listener from the server configurations
